#1 - Added Spring Boot starter.
Add dedicated module to simplify Spring Boot integration. Pulls in the JPA based registry and the Jackson based serializer. Activates auto-configuration of the events package. Removed explicit @EnableJpaRepositories in case we're running in a Boot application as it disables auto-configuration of JPA for the rest of the application.

Switched to dependency management of Spring Boot.