#3 - Completion interceptor is now ordered behind the @Async one.
CompletionRegisteringMethodInterceptor now implements Ordered and selects an order to be placed behind an @Async one. That makes sure that the completion is part of the asynchronous execution and only gets written if the asynchronous invocation succeeds.

Simplified the interceptor setup to avoid superfluous instance creation. The proxy creation now also detects existing proxies and adds the interceptor to the current advice list in that case. We now also properly prevent duplicate interceptor registration.