#4 - Make @ModuleTest now also tweak Spring Boot's auto-configuration and entity scan packages.
We now manipulate the packages used for auto-configuration and entity scanning to limit them to only use the package of the class annotated with @ModuleTest. This needs to be refined to actually use the module's base package to support @ModuleTest to be used in sub-packages but is fine for now.