#2 - Added support for more complex package structures and named module interfaces.
We now have an @NamedModule annotation that can be used to customize which packages of a module are supposed to be exposed as dependency targets for other modules.

Tweaked the way we look up module packages as previously we relied on a simple component scan which only works if the module base package contains at least one file. We now scan all packages and extract all direct sub packages even from nested ones.

Added info level logging about which modules are included in a module test.