Improved Spring dependency handling.
We now distinguish between the general dependency analysis and the bootstrap dependencies that are required to run an individual test. The latter only includes Spring component dependencies towards other modules but e.g. exclude plain type dependencies that establish domain type dependencies, event listeners etc.

The cycle prevention check is now established on all root packages. Spring components declared through @Bean methods are now considered, too.

Modules.of(…) now caches instances by Modulith base type to avoid repeated classpath scans.

Replaced ModuleContextLoader with a ModuleContextCustomizerFactory as the ContextCustomizer created by the factory are automatically included in the key for the context caching. The ModuleContextCustomizer's equals(…) and hashCode() are implemented based on ModuleTestExecution which in turn are based on the module base package and @ModuleTest annotation. That should trigger ApplicationContext reuse for all test cases for a specific module and the actual @ModuleTest setup.