#26 - More options to create individual diagrams in Documenter.
Documenter now exposes methods that take an Options argument to control the output of the rendering of either all modules or individual ones. Dependency relationships are now deduplicated, which means that the high level relationships (component dependency, event listeners) which implicitly imply the plain type dependencies are favored and the latter are not rendered by default. They still show up in case the high level relationships are excluded via the Options parameter.