#74 - Support for named interfaces based on annotated types.
Previously, named interfaces could only be specified on the package level using an annotation. That annotation can now be used on types as well so that named interfaces can be formed from any arbitrary set of classes. The annotation also allows to define multiple names so that a type can be made belong to multiple interfaces.