#98 - Improve event capture during test cases.
We now avoid registering an additional event listener for each test method that takes a PublishedEvents parameter. We register one initially and then for each method invocation we register a PublishedEvents instance to and bind it to the current thread. We unregister that instance after the test methods has been concluded. Using a thread bound delegate makes sure that this also works if tests are executed concurrently.

Original pull request: #99.