#101 - Added support for writing a module canvas.
We now allow to create module canvases for each module in Asciidoctor formats listing the following attributes:

- Base package
- Spring components contained in the module, including the interfaces exposed contained in the module
- Aggregates defined by the module (entities for which there's a repository available)
- Events listened to
- Events published

If a Javadoc base URI os provided in the call to Documenter.toModuleCanvas(…) all type renderings will be wrapped into links pointing to the published Javadoc. Published event types will need to be annotated with the @Event annotation newly introduced in the API module.