#137 - Introduce ModuleDetectionStrategy SPI.
We now allow the customization of which packages are considered module base packages via the ModuleDetectionStrategy interface. Default implementations are exposed as factory methods on the interface itself. Custom implementations can be registered via an entry for org.moduliths.model.ModuleDetectionStrategy in a META-INF/spring.factories file on the classpath.