#121 - Rearrange configuration arrangement for JpaEventPublicationRegistry.
We now use explicit bean factory method parameters for the dependencies for JpaEventPublicationRegistry to be created. This is needed to let Spring Framework detect the logical bean dependency between the registry and the repository to properly shut them down in the correct order.

The previously used arrangement via configuration class constructor arguments hides that dependency and potentially already shuts down the repository or downstream beans which then will cause the shutdown hook in JpaEventPublicationRegistry itself to fail as it still needs to trigger a repository query in that.