#171 - Allow Modulith events JPA module be used with DevTools.
Added necessary metadata to make sure the package protected  JpaEventPublicationRepository can be bootstrapped when Spring Boot's DevTools are in place. With the declaration present, the JAR's code is placed into the reload classloader so that the proxy class created is held in that one, too, and thus is allowed to refer to the package protected interface.