#212 - Better output folder defaults in Documenter.
We now pick a suitable output folder to write the generated files to. We stick to the previous default if we find a pom.xml in the project root, indicating a Maven project. If we don't find that file, we assume a Gradle project and tweak the default to build/moduliths-docs.

If users want to override this, they can now also use the ….withOutputFolder(…) to explicitly customize this.