#201 - Fix republication of outstanding event completions on restart.
We now directly call ….processEvent(…) on the TransactionalApplicationListener to immediately invoke it rather than ….onApplicationEvent(…), which would attach the actual publication to the commit phase of a transaction (which currently isn't even in progress).

We now also avoid unconditionally marking the publication as completed as the exception handling is already done in CompletionRegisteringMethodInterceptor. It also takes care of logging so that we can remove that as well.

Additional tests that make sure the expected listeners are invoked another time but unrelated ones are not and we still end up with the same amount of incomplete publications.