#229 - Add starters for easier consumption.
Introduced starter JARs for JPA, JPA (Jakarta) and Test to make it easier to refer to Moduliths. The JPA starter contains Moduliths API and core modules, Moments and the events subsystem using JPA. The Jakarta flavour is basically the same but using JPA in its JakartaEE variant. The test starter includes both the test support as well as the documentation bits.