#249 - Avoid non-component dependencies in bootstrap ones.
Module.getBootstrapDependencies(…) previously returned all modules a module would depend on, including ones that do not need to be bootstrapped as the dependency might only be established by an entity or event listener.

A few renames in Module's method names to make clear what kind of dependencies and packages are returned.