GH-254 - Module test autoconfiguration is now ordered in highest priority.
To make sure the package augmentation is applied before any other autoconfiguration picks it up. Also, we now stick to the String[] type of the constructor argument values of the modified bean definition as Boot 3.0 introduced code that inspects those values and thus can't deal with us changing them to a list.